♻️ refactor : 엘라스틱서치 패키지를 jobopening 외부로 분리
