+++ src/main/java/com/project/cheerha/domain/elasticsearch/dto/response/ReadJobOpeningElasticResponseDto.java
@@ -1,4 +1,4 @@
-package com.project.cheerha.domain.jobopening.elasticsearch.dto.response;
+package com.project.cheerha.domain.elasticsearch.dto.response;
 
 import java.time.ZonedDateTime;
 import java.util.List;
