+++ src/main/java/com/project/cheerha/domain/elasticsearch/entity/JobOpeningDocument.java
@@ -1,7 +1,7 @@
-package com.project.cheerha.domain.jobopening.elasticsearch.entity;
+package com.project.cheerha.domain.elasticsearch.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.project.cheerha.domain.jobopening.elasticsearch.IndexName;
+import com.project.cheerha.domain.elasticsearch.IndexName;
 import com.project.cheerha.domain.jobopening.entity.JobOpening;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
