+++ src/main/java/com/project/cheerha/domain/elasticsearch/repository/JobOpeningDocumentRepository.java
@@ -1,6 +1,6 @@
-package com.project.cheerha.domain.jobopening.elasticsearch.repository;
+package com.project.cheerha.domain.elasticsearch.repository;
 
-import com.project.cheerha.domain.jobopening.elasticsearch.entity.JobOpeningDocument;
+import com.project.cheerha.domain.elasticsearch.entity.JobOpeningDocument;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 public interface JobOpeningDocumentRepository extends ElasticsearchRepository<JobOpeningDocument, String> {
