+++ src/main/java/com/project/cheerha/domain/jobopening/service/JobOpeningService.java
@@ -3,7 +3,7 @@ package com.project.cheerha.domain.jobopening.service;
 import com.project.cheerha.domain.history.service.HistoryService;
 import com.project.cheerha.domain.jobopening.dto.request.ReadJobOpeningRequestDto;
 import com.project.cheerha.domain.jobopening.dto.response.ReadJobOpeningResponseDto;
-import com.project.cheerha.domain.jobopening.elasticsearch.IndexName;
+import com.project.cheerha.domain.elasticsearch.IndexName;
 import com.project.cheerha.domain.jobopening.entity.JobOpening;
 import com.project.cheerha.domain.jobopening.repository.JobOpeningRepository;
 import com.project.cheerha.domain.user.entity.User;
